requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.0
opencv-python==4.1.0.25
protobuf==3.9.1
requests2==2.16.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.21.1
Werkzeug==0.15.5
wrapt==1.11.2
